Update CI to Qt 6.3.1 and minor code changes (#36)
* clang-cl: disable pre-c99 compatibility warning

* strict types for enums

* update to qt 6.3.1

+++ src/bootentry.cpp
@@ -104,7 +104,7 @@ auto BootEntry::toJSON() const -> QJsonObject
 {
     QJsonObject load_option;
     load_option["description"] = description;
-    load_option["optional_data_format"] = optional_data_format;
+    load_option["optional_data_format"] = static_cast<int>(optional_data_format);
     load_option["optional_data"] = optional_data;
     load_option["attributes"] = static_cast<int>(attributes);
     load_option["efi_attributes"] = static_cast<int>(efi_attributes);
